requirements.txt
Django
requests
SQLAlchemy
sqlparse
psycopg2-binary
urllib3
Pillow
djangorestframework
